REFACTORED SQLT PARSER TO CALL $schema->dep_ordered_sources
Pulled out code in the SQLT Parser that was calling private method
$schema->_resolve_deps (with tangled legacy API) and converted to call
the new schema method 'dep_ordered_sources' which handles figuring out
the ordering of sources needed for deploy. This now completes the
logical consolidation/deduplication work (and then some) for the
topic/factor-out-source-tree branch.

Note that the deploy order has changed from what it used to be for many
schemas (such as DBICTest). The previous order was only "good enough"
for a safe deploy order in tested schemas. The new order is more
accurate as it considers the actual dependencies and not just the number
of dependencies.

Also added new arg opt 'limit_deps' to source_tree/dep_ordered_sources
see POD.

Also added more tests